+++ src/Modules/ElementUtils/object.js
@@ -1,0 +1,13 @@
+import isEmpty from 'lodash/isEmpty';
+
+function hasProperties(element) {
+  if (!element) {
+    return false;
+  }
+
+  return !isEmpty(element.content);
+}
+
+export {
+  hasProperties,
+};
