+++ src/Modules/ElementUtils/structuredElement.js
@@ -1,0 +1,30 @@
+import {MissingCacheObjectException} from '../../Exceptions/MissingCacheObject';
+
+function constainsStructuredElement(element) {
+  if (!element) {
+    return false;
+  }
+
+  if (!element.cache) {
+    throw new MissingCacheObjectException(element);
+  }
+
+  return element.cache.containsStructuredElement;
+}
+
+function isStructured(element) {
+  if (!element) {
+    return false;
+  }
+
+  if (!element.cache) {
+    throw new MissingCacheObjectException(element);
+  }
+
+  return element.cache.isStructured;
+}
+
+export {
+  containsStructuredElement,
+  isStructured,
+};
