+++ src/Modules/ElementUtils/value.js
@@ -1,0 +1,31 @@
+import {isObjectOrArray, isMember} from './type';
+
+function hasValue(element) {
+  if (!element) {
+    return false;
+  }
+
+  if (isObjectOrArray(element.element)) {
+    return false;
+  }
+
+  if (!element.content) {
+    return false;
+  }
+
+  if (isMember(element.element)) {
+    if (element.content.value && isObjectOrArray(element.content.value.element)) {
+      return false;
+    }
+
+    if (!element.content.value.content) {
+      return false;
+    }
+  }
+
+  return true;
+}
+
+export {
+  hasValue,
+};
