+++ src/Resources/types.js
@@ -8,4 +8,6 @@ const TYPES = {
   SELECT: 'select',
 };
 
-export default TYPES;
+export {
+  TYPES,
+};
